# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Search,Insert,Delete value in BST",
        "title": "BST",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "description": "welcome page",
                "tags": [
                    "main"
                ],
                "summary": "MainPage",
                "responses": {}
            }
        },
        "/delete": {
            "get": {
                "description": "remove value from BST",
                "tags": [
                    "remove"
                ],
                "summary": "Remove",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "int",
                        "name": "remove",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/insert": {
            "get": {
                "description": "insert value in BST",
                "tags": [
                    "insert"
                ],
                "summary": "Insert",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "int",
                        "name": "insert",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/search": {
            "get": {
                "description": "lookup value in BST",
                "tags": [
                    "search"
                ],
                "summary": "Search",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "int",
                        "name": "search",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    }
}